Make TagFilter a long, only because a lot of modern languages treat int as 64bit

+++ Shoko.Server/Utilities/TagFilter.cs
@@ -379,7 +379,7 @@ namespace Shoko.Server
         };
 
         [Flags]
-        public enum Filter : int
+        public enum Filter : long
         {
             AnidbInternal = 1 << 0,
             ArtStyle      = 1 << 1,
